requirements.txt
# Core
streamlit>=1.28.0
python-dotenv>=1.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ML Explainability
shap>=0.42.0
lime>=0.2.0.1

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0

# PDF Generation
reportlab>=4.0.0

# Database
sqlite3  # Built-in with Python

# API
openai>=1.0.0  # For AI copilot feature

# Configuration
pyyaml>=6.0.0

# Optional: Development
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0
